+++ src/lang-server.ts
@@ -3,11 +3,10 @@ import {
   LanguageClientOptions,
   Executable,
 } from 'vscode-languageclient/node'
+import * as which from 'which'
 import * as vscode from 'vscode'
 import {
-  getBinPath,
-  FORTRAN_FREE_FORM_ID,
-  promptForMissingTool,
+  FORTRAN_DOCUMENT_SELECTOR,
 } from './lib/helper'
 import { LANG_SERVER_TOOL_ID } from './lib/tools'
 
@@ -18,13 +17,13 @@ export class FortranLangServer {
     let langServerFlags: string[] = config.get('languageServerFlags', [])
 
     const serverOptions: Executable = {
-      command: getBinPath(LANG_SERVER_TOOL_ID),
+      command: which.sync(LANG_SERVER_TOOL_ID),
       args: [...langServerFlags],
       options: {},
     }
 
     const clientOptions: LanguageClientOptions = {
-      documentSelector: [FORTRAN_FREE_FORM_ID],
+      documentSelector: FORTRAN_DOCUMENT_SELECTOR,
     }
 
     this.c = new LanguageClient(
@@ -48,23 +47,3 @@ export class FortranLangServer {
     return capabilities
   }
 }
-
-export function checkForLangServer(config) {
-  const useLangServer = false //config.get('useLanguageServer')
-  if (!useLangServer) return false
-  if (process.platform === 'win32') {
-    vscode.window.showInformationMessage(
-      'The Fortran language server is not supported on Windows yet.'
-    )
-    return false
-  }
-  let langServerAvailable = getBinPath(LANG_SERVER_TOOL_ID)
-  if (!langServerAvailable) {
-    promptForMissingTool(LANG_SERVER_TOOL_ID)
-    vscode.window.showInformationMessage(
-      'Reload VS Code window after installing the Fortran language server'
-    )
-  }
-  return true
-}
-